+++ fw_al1mb1/app_header/app_header.h
@@ -49,6 +49,9 @@ typedef struct __attribute__((packed)) {
 #define APP_HEADER_MAGIC    0xDEADBEEF
 #define APP_VERSION         0x00010100  /* Version 1.1.0 */
 
+/* Application header instance (defined in app_header.c, placed at 0x08004000) */
+extern const app_header_t app_header;
+
 /* USB VID/PID - customize these or use defaults (STMicroelectronics DFU) */
 #ifndef USB_VID
 #define USB_VID             0x0483      /* STMicroelectronics */